chore: merge new tests into existing test suite
